requirements.txt
flask==3.0.2
groq==0.19.0
python-dotenv==1.0.1
flask-wtf==1.2.1
PyPDF2==3.0.1
reportlab==4.1.0
python-magic==0.4.27
python-docx==1.1.0
markdown==3.5.2